+++ sdk/search/Azure.Search.Documents/src/Models/SearchModelFactory.cs
@@ -590,8 +590,7 @@ namespace Azure.Search.Documents.Models
 
         /// <summary> Initializes a new instance of DocumentDebugInfo. </summary>
         /// <param name="vectors"> Contains debugging information specific to vector and hybrid search. </param>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public static DocumentDebugInfo DocumentDebugInfo(VectorsDebugInfo vectors = null)
+        internal static DocumentDebugInfo DocumentDebugInfo(VectorsDebugInfo vectors = null)
         {
             return new DocumentDebugInfo(vectors, additionalBinaryDataProperties: null);
         }